+++ src/test/groovy/com/codebrig/arthur/observe/structure/filter/operator/relational/RelationalOperatorFilterTest.groovy
@@ -2,12 +2,8 @@ package com.codebrig.arthur.observe.structure.filter.operator.relational
 
 import com.codebrig.arthur.ArthurTest
 import com.codebrig.arthur.SourceLanguage
-import com.codebrig.arthur.SourceNode
 import com.codebrig.arthur.observe.structure.filter.FunctionFilter
-import com.codebrig.arthur.observe.structure.filter.InternalRoleFilter
 import com.codebrig.arthur.observe.structure.filter.MultiFilter
-import com.codebrig.arthur.observe.structure.filter.RoleFilter
-import com.codebrig.arthur.observe.structure.filter.TypeFilter
 import com.codebrig.arthur.observe.structure.filter.operator.relational.define.DeclareVariableOperatorFilter
 import gopkg.in.bblfsh.sdk.v1.protocol.generated.Encoding
 import org.junit.Test
@@ -61,47 +57,13 @@ class RelationalOperatorFilterTest extends ArthurTest {
         new FunctionFilter().getFilteredNodes(language, resp.uast).each {
             MultiFilter.matchAll(new RelationalOperatorFilter()).reject(new DeclareVariableOperatorFilter())
                     .getFilteredNodes(it).each {
-                def left = null
-                def right = null
-                def matchedLeft = MultiFilter.matchAny(
-                        new RoleFilter("LEFT"),
-                        new InternalRoleFilter("Left")
-                ).getFilteredNodes(it.children)
-                left = (matchedLeft.hasNext()) ? matchedLeft.next() : left
-                def matchedRight = MultiFilter.matchAny(
-                        new RoleFilter("RIGHT"),
-                        new InternalRoleFilter("Right")
-                ).getFilteredNodes(it.children)
-                right = (matchedRight.hasNext()) ? matchedRight.next() : right
-                left = (left == null) ? getPropOperand1(it.children) : left
-                right = (right == null) ? getPropOperand2(it.children) : right
-                assertNotNull(left)
-                assertNotNull(right)
+                assertNotNull(RelationalOperatorFilter.getLeftOperand(it))
+                assertNotNull(RelationalOperatorFilter.getRightOperand(it))
                 foundLeftOperands = true
                 foundRightOperands = true
             }
         }
         assertTrue(foundLeftOperands)
         assertTrue(foundRightOperands)
-    }
-
-    private static SourceNode getPropOperand1(Iterator<SourceNode> children) {
-        def node = null
-        new TypeFilter("CPPASTBinaryExpression").getFilteredNodes(children).each {
-            new InternalRoleFilter("Prop_Operand1").getFilteredNodes(it).each {
-                node = it
-            }
-        }
-        return node
-    }
-
-    private static SourceNode getPropOperand2(Iterator<SourceNode> children) {
-        def node = null
-        new TypeFilter("CPPASTBinaryExpression").getFilteredNodes(children).each {
-            new InternalRoleFilter("Prop_Operand2").getFilteredNodes(it).each {
-                node = it
-            }
-        }
-        return node
     }
 }
